1) File.h: File and derived classes now have a member variable m_state, which keeps track of Errno. The current implementation guarantees that m_state is update ONLY if it is Errno::good. That is, new error numbers do not overwrite pre-existing error numbers. 2) ImageFile.h, Header.h, AbstractHeader.h: Choosing the type of header for the ImageFile class can be a runtime decision now. This is much more flexible. The current implementation uses an opaque pointer held by Header.h and all type headers are children of AbstractHeader. 3) IO.h: IO::Layout is now an enum class. This is cleaner and simpler to use.

+++ src/noa/managers/InputManager.h
@@ -48,11 +48,11 @@ namespace Noa {
          *  - 5 @c docstring     : docstring displayed with the @c "--help" command.
          */
         struct OptionUsage {
-            static constexpr u_int8_t long_name{0U};
-            static constexpr u_int8_t short_name{1U};
-            static constexpr u_int8_t type{2U};
-            static constexpr u_int8_t default_value{3U};
-            static constexpr u_int8_t docstring{4U};
+            static constexpr uint32_t long_name{0U};
+            static constexpr uint32_t short_name{1U};
+            static constexpr uint32_t type{2U};
+            static constexpr uint32_t default_value{3U};
+            static constexpr uint32_t docstring{4U};
         };
 
         const std::string m_usage_header = fmt::format(
